Offset calculations can now use sequences with mismatched number of crossings btwn channels

+++ Scope_Interfacing_Code/scope_connect_test.py
@@ -2,48 +2,64 @@ import win32com.client #imports the pywin32 library
 import scope_script_MDP as ss
 from MAUI import MAUI
 import matplotlib.pyplot as plt
-import time
+import shutil
 import numpy as np
 import os, gc
 
 if __name__=="__main__":
     
-    ref_array_list = []
-    chip_array_list = []
-    offset_vals_list = []
+    delete_prev_data = True # Delete data from previous experiments
+
+    N = 100 # Number of acquisitions per sequence
+    num_samples = int(1e4) # Number of samples per acquisition segment in the sequence
+    div_time = 50e-9 # There are 10 divisons per acquisition
     
-    N = 20
-    num_samples = int(1e4) 
-    div_time = 50-9 # There are 10 divisons per acquisition
+    hold_time = 1e-8 # Chip falling edge must occur within this many seconds after ref rising edge. Should be less than AWG signal high width
     
-    num_loops = 2
+    num_loops = 50 # Number of sequences 
+
+    # Voltage thresholds for reference and chip signals
     ref_thresh = .08
     chip_thresh = 0.00
+
+    # Number of initial samples to discard when processing data
+    # There's consistently a weird signal spike during the first acquisition, so we discard that data
     clip = num_samples
-    
+
     # Make appropriate (sub)directories for storing data from each loop
     save_dir = "C:\\LeCroy\\ScopeData"
     save_dir_ref = save_dir + "\\ReferenceWaveforms"
     save_dir_chip = save_dir + "\\ChipWaveforms"
     save_dir_offset = save_dir +"\\OffsetVals"
 
+    # Create final files that will hold data from all sequences
+    combined_offset_file = os.path.join(save_dir, "offset_values_all.txt")
+    # combined_ref_data_file = os.path.join(save_dir, "ref_data_net.npy")
+    # combined_chip_data_file = os.path.join(save_dir, "chip_data_net.npy")
+
+    if delete_prev_data:
+        
+        # Delete folders with data files
+        if os.path.exists(save_dir_ref):
+            shutil.rmtree(save_dir_ref)
+        if os.path.exists(save_dir_chip):
+            shutil.rmtree(save_dir_chip)
+        if os.path.exists(save_dir_offset):
+            shutil.rmtree(save_dir_offset)
+
+        # Delete cumulative files
+        if os.path.exists(combined_offset_file):
+            os.remove(combined_offset_file)
+        # if os.path.exists(combined_ref_data_file):
+        #     os.remove(combined_ref_data_file)
+        # if os.path.exists(combined_chip_data_file):
+        #     os.remove(combined_chip_data_file)
+
+
     os.makedirs(save_dir, exist_ok=True)
     os.makedirs(save_dir_ref, exist_ok=True)
     os.makedirs(save_dir_chip, exist_ok=True)
     os.makedirs(save_dir_offset, exist_ok=True)
-
-    # Create final files that will hold "all" the data
-    combined_offset_file = os.path.join(save_dir, "offset_values_all.txt")
-    combined_ref_data_file = os.path.join(save_dir, "ref_data_net.npy")
-    combined_chip_data_file = os.path.join(save_dir, "chip_data_net.npy")
-    
-    # Remove previous data if it's still there
-    if os.path.exists(combined_offset_file):
-        os.remove(combined_offset_file)
-    if os.path.exists(combined_ref_data_file):
-        os.remove(combined_ref_data_file)
-    if os.path.exists(combined_chip_data_file):
-        os.remove(combined_chip_data_file)
 
     with MAUI() as c:
         loop = 0
@@ -54,8 +70,9 @@ if __name__=="__main__":
             c.reset()
 
             # Set appropriate voltage scales and time divisions for acquisition
-            c.set_vertical_scale("C1", .05)# For AWG this should be 20 mV (.020)
+            c.set_vertical_scale("C1", .05) # For AWG this should be 20 mV (.020)
             c.set_vertical_scale("C2", .22)
+            c.set_timebase(div_time)
 
             c.idn() # Needed for code to run for some reason
 
@@ -72,7 +89,8 @@ if __name__=="__main__":
                                                         N=N,
                                                         num_samples=num_samples, 
                                                         ref_channel="C1", ref_edge_slope="POS", ref_thresh=ref_thresh,
-                                                        chip_channel="C2", chip_edge_slope="NEG", chip_thresh=chip_thresh)
+                                                        chip_channel="C2", chip_edge_slope="NEG", chip_thresh=chip_thresh,
+                                                        hold_time=hold_time)
             print(f"\tData acquired")
             # Add approprite offset to time data
             ref_data[0] = ref_data[0] + time_this_loop
@@ -85,13 +103,15 @@ if __name__=="__main__":
                                     chip_data,
                                     ref_threshold=ref_thresh,
                                     chip_threshold=chip_thresh,
-                                    clip=clip)
+                                    clip=clip,
+                                    mismatch_handling=True,
+                                    num_samples=num_samples)
                 
                 print(f"\tOffsets calculated")
                 
                 # Save wave data to files specific to this loop
-                np.save(ref_data_file_i, ref_data.T)
-                np.save(chip_data_file_i, chip_data.T)
+                np.save(ref_data_file_i, ref_data)
+                np.save(chip_data_file_i, chip_data)
                 
                 # Just so we don't overflow memory; just for testing
                 # os.remove(ref_data_file_i)
@@ -107,7 +127,8 @@ if __name__=="__main__":
                 loop += 1
             
             # If data isn't properly formatted, discard it and move on to the next loop
-            except ValueError:
+            except ValueError as e:
+                print(str(e))
                 print(f"\tDiscarding problematic waveforms from this loop")
             
                 del ref_data, chip_data
@@ -129,11 +150,9 @@ if __name__=="__main__":
         os.rmdir(save_dir_offset)
         
         offset_vals_all = np.loadtxt(combined_offset_file)
-
-
         mean_val, std_val = ss.calculate_mean_and_std(offset_vals_all)
         
         print(f"\nAverage offset btwn edges: {mean_val}")
         print(f"Stdv of offset time: {std_val}")
 
-        hist, bin_edges = ss.make_histogram_and_gaussian(offset_vals_all, hist_bins=40)
+        hist, bin_edges = ss.make_histogram_and_gaussian(offset_vals_all, hist_bins=40, stdv_cutoff=10)
